refactor(platform): wip job detail page cleanup
