+++ turbo/apps/platform/src/signals/agent.ts
@@ -59,13 +59,13 @@ export const currentAgentId$ = computed((get) => {
   return typeof id === "string" ? id : null;
 });
 
-// ---------------------------------------------------------------------------
-// Identity — sidebar agent (user-selected, falls back to default)
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// Agent lists
-// ---------------------------------------------------------------------------
+export const currentAgent$ = computed((get) => {
+  const agentId = get(currentAgentId$);
+  if (!agentId) {
+    return null;
+  }
+  return get(agentById(agentId));
+});
 
 const internalReloadAgents$ = state(0);
